Activate KDA Galaxy performance calibration gate
